[master] explicit casting
- 'return': conversion from 'size_t' to 'int', possible loss of data
- connection maximum is less than socket connection

+++ server/SesNotSvc/BroadcastTcpServer.cpp
@@ -200,7 +200,7 @@ private:
 		::EnterCriticalSection(&criticalSection);
 		count = this->subscribers.size();
 		::LeaveCriticalSection(&criticalSection);
-		return count;
+		return (int)count;
 	}
 };
